% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{align_sequences}
\alias{align_sequences}
\title{Align Sequences}
\usage{
align_sequences(n_sample.scalar, n_reference.scalar, sample.shortreadq,
  reference.filepath, reference.shortread)
}
\arguments{
\item{n_sample.scalar}{The number of sample sequences}

\item{n_reference.scalar}{The number of reference sequences}

\item{sample.shortreadq}{A ShortRead object containing the sample sequences}

\item{reference.filepath}{Path to file for the reference sequences}

\item{reference.shortread}{A ShortRead object containing the reference sequences}
}
\value{
A MsaDNAMultipleAlignment object containing the aligned sequences
}
\description{
This function aligns sequences with ClustalOmega. It combines sample and reference sequences together before alignment.
If reference sequences are not available, the sample sequences alone are aligned.
}
