% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{aligned_reference_sequences}
\alias{aligned_reference_sequences}
\title{Select Reference Sequences from Alignment}
\usage{
aligned_reference_sequences(alignment.character, n_sample.scalar,
  n_reference.scalar)
}
\arguments{
\item{alignment.character}{A matrix or vector containing the alignment}

\item{n_sample.scalar}{The number of sample sequences}

\item{n_reference.scalar}{The number of reference sequences}
}
\value{
A character vector containing the reference sequences (aligned)
}
\description{
This function selects the reference sequences from the alignment.  Sample and reference
sequences are aligned together, and reference sequences must be seperated.
}
