% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{aligned_sample_sequences}
\alias{aligned_sample_sequences}
\title{Select Sample Sequences from Alignment}
\usage{
aligned_sample_sequences(alignment.character, n_sample.scalar)
}
\arguments{
\item{alignment.character}{A matrix or character vector containing the alignment}

\item{n_sample.scalar}{The number of sample sequences}
}
\value{
A character vector containing samples sequences (aligned)
}
\description{
This function selects the sample sequences from the alignment.  Sample and reference
sequences are aligned together, and sample sequences must be seperated.
}
