% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{alignment_character}
\alias{alignment_character}
\title{Convert Alignment to Character Vector}
\usage{
alignment_character(alignment.msa)
}
\arguments{
\item{alignment.msa}{A MsaDNAMultipleAlignment object containing the aligned sequences}
}
\value{
A character vector containing the alignment
}
\description{
This function converts the an alignment of the MsaDNAMultipleAlignment class to character
vector.  The alignment is converted to a matrix before being converted to a vector.
}
