% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{alignment_width}
\alias{alignment_width}
\title{Determine Alignment Width}
\usage{
alignment_width(alignment.msa)
}
\arguments{
\item{alignment.msa}{A MsaDNAMultipleAlignment object containing the aligned sequences}
}
\value{
The number of columns in the alignment
}
\description{
This function determines how many columns are in an alignment.  The alignment
is converted to a matrix before determining the number of columns.
}
