% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{distance_fractional}
\alias{distance_fractional}
\title{Calculate Fractional Distance}
\usage{
distance_fractional(distances_observed.matrix, n_align.scalar,
  n_shared_indels.matrix)
}
\arguments{
\item{distances_observed.matrix}{A matrix containing Hamming distances}

\item{n_align.scalar}{The number of columns in the alignment}

\item{n_sample.scalar}{The number of sample sequences}
}
\value{
A matrix containing the fractional distances
}
\description{
This function calculates the fractional distance (the fraction of letters that differ) between pairs of sequences.
The number of shared indels is subtracted from the number of aligned nucleotide positions.
The distances are placed in a matrix.  The upper matrix triangle is copied from the lower triangle.
}
