% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{distance_fractional_estimated}
\alias{distance_fractional_estimated}
\title{Calculate Fractional Distance (Estimated)}
\usage{
distance_fractional_estimated(alignment_sample.matrix, n_align.scalar,
  n_sample.scalar, errorrates_estimated_aligned.matrix)
}
\arguments{
\item{alignment_sample.matrix}{A matrix containing Hamming distances}

\item{n_align.scalar}{The number of columns in the alignment}

\item{n_sample.scalar}{The number of sample sequences}

\item{errorrates_estimated_aligned.matrix}{A matrix containing the error probabilities (aligned)}
}
\value{
A matrix containing the fractional distances
}
\description{
This function calculates the fractional distance between pairs of sequences, but it
corrects the distance using eq. 1 in Hackmann (2019. Bioinformatics. In press.)
This function is distinct from distance_fractional.
}
