% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{distance_Hamming}
\alias{distance_Hamming}
\title{Calculate Hamming Distance}
\usage{
distance_Hamming(n_sample.scalar, alignment_sample.character)
}
\arguments{
\item{n_sample.scalar}{The number of sample sequences}

\item{alignment_sample.character}{A character vector containing the sample sequences (aligned)}
}
\value{
A matrix containing Hamming distances
}
\description{
This function calculates the Hamming distance (the number of letters that differ) between pairs of sequences.
}
