% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{Distanced}
\alias{Distanced}
\title{$$$$$$$$$$$
$DISTANCED$
$$$$$$$$$$$}
\usage{
Distanced(sample.filepath, reference.filepath, n_sample_max.scalar,
  replace_ambiguous_letters.character, random_seed.character)
}
\arguments{
\item{sample.filepath}{Path to file for sample sequences}

\item{reference.filepath}{Path to file for reference sequences}

\item{n_sample_max.scalar}{The number of reads to be subsampled}

\item{replace_ambiguous_letters.character}{Should Ns be replaced? (TRUE = Yes, FALSE = No)}

\item{random_seed.character}{Should the seed be set to 1? (TRUE = No, FALSE = Yes)}
}
\value{
A dataframe containing the MPD values
}
\description{
Distanced calculates the mean pairwise distance (MPD)
of nucleic acid sequences.  It requires a set of sample
sequences in fastq or fasta format.  A set of reference
sequences is optional.  After loading the sequences,
Distanced takes a random subsample of sample sequences.
It aligns sample and reference sequences against each other.
It calculates three different values of MPD (uncorrected, estimated,
actual) and returns them in a dataframe.
}
