% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{error_probabilities_aligned}
\alias{error_probabilities_aligned}
\title{Calculate Error Probabilities}
\usage{
error_probabilities_aligned(n_align.scalar, n_sample.scalar,
  sample.shortreadq, alignment_sample.matrix)
}
\arguments{
\item{n_align.scalar}{The number of columns in the alignment}

\item{n_sample.scalar}{The number of sample sequences}

\item{sample.shortreadq}{A ShortRead object containing the sample sequences}

\item{alignment_sample.matrix}{A matrix containing Hamming distances}
}
\value{
A matrix containing the error probabilities (aligned)
}
\description{
This function calculates error probabilities for each letter in the sample
sequence.  Error probabilties are the fractional probability that a letter is wrong.
They are calculated from the quality scores.  The quality scores are from unaligned sequences.
The function also brings the error probabilities into the same alignment as the sequence letters.
}
