% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{load_reference_sequences}
\alias{load_reference_sequences}
\title{Load Reference Sequences}
\usage{
load_reference_sequences(reference.filepath)
}
\arguments{
\item{reference.filepath}{Path to  file for reference sequences}

\item{sample_filetype_FASTQ.character}{Sample file type (TRUE = fastq, FALSE = fasta)}
}
\value{
A ShortRead object containing sample sequences
}
\description{
This function loads a fasta file containing the reference sequences, if the user provides them.
}
