% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{load_sample_sequences}
\alias{load_sample_sequences}
\title{Load Sample Sequences}
\usage{
load_sample_sequences(sample.filepath, sample_filetype_FASTQ.character)
}
\arguments{
\item{sample.filepath}{Path to file for sample sequences}

\item{sample_filetype_FASTQ.character}{Sample file type (TRUE = fastq, FALSE = fasta)}
}
\value{
A ShortRead object containing sample sequences
}
\description{
This function loads a fasta or fastq file containing the sample sequences.
Sample sequences should not contain chimeras or primers.
}
