% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{match_reference_to_sample_sequence}
\alias{match_reference_to_sample_sequence}
\title{Find Matches Between Sample and Reference Sequences}
\usage{
match_reference_to_sample_sequence(alignment_sample.character,
  alignment_reference.character)
}
\arguments{
\item{alignment_sample.character}{A character vector containing the sample sequences (aligned)}

\item{alignment_reference.character}{A character vector containing the reference sequences (aligned)}
}
\value{
A matrix containing the indices of the matching reference sequences
}
\description{
This function finds the reference sequence that most closely
matches a given sample sequence.  The matching sequene is that
which has the lowest Hamming distance.  It return the index
of the reference sequence, rather than the sequence itself.
}
