% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{matching_reference_sequence}
\alias{matching_reference_sequence}
\title{Get Matching Reference Sequences}
\usage{
matching_reference_sequence(alignment_reference.character,
  reference_index.matrix)
}
\arguments{
\item{alignment_reference.character}{A matrix or character vector containing the reference sequences (aligned)}

\item{reference_index.matrix}{A matrix containing the indices of the matching reference sequences}
}
\value{
A matrix or character vector containing matching reference sequences
}
\description{
This function gets the reference sequence that most closely
matches a given sample sequence.  It returns the sequence itself.
}
