% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{MPD_calculation}
\alias{MPD_calculation}
\title{Calculate Mean Pairwise Distance}
\usage{
MPD_calculation(distances_observed.matrix, n_sample.scalar)
}
\arguments{
\item{distances_observed.matrix}{A matrix containing the fractional distances}

\item{n_sample.scalar}{The number of sample sequences}
}
\value{
The mean pairwise distance
}
\description{
This function calculates mean pairwise distance (MPD).
This is the average of the fractional distance.
Before calculating the mean, the dractional distances are taken from the orignal matrix,
placed in a vector, #' and then diagonal elements from the matrix.
}
