% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{MPD_summary}
\alias{MPD_summary}
\title{Summarize Mean Pairwise Distance Values}
\usage{
MPD_summary(MPD_observed.scalar, MPD_original_estimated.scalar,
  MPD_original_actual.scalar)
}
\arguments{
\item{MPD_observed.scalar}{The uncorrected MPD}

\item{MPD_original_estimated.scalar}{The estimated MPD}

\item{MPD_original_actual.scalar}{The actual MPD}
}
\value{
A dataframe containing the uncorrected, estimated, and actual MPD
}
\description{
This function summarizes uncorrected, estimated, and actual
mean pairwise distance (MPD) in a dataframe.
}
