% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{n_reference_sequences}
\alias{n_reference_sequences}
\title{Detetermine Number of Reference Sequences}
\usage{
n_reference_sequences(reference.shortread, reference.filepath)
}
\arguments{
\item{reference.filepath}{Path to file for reference sequences}

\item{sample.shortreadq}{A ShortRead object containing the reference sequences}
}
\value{
The number of sequences
}
\description{
This function determines the number of reference sequences.
}
