% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{n_sample_sequences}
\alias{n_sample_sequences}
\title{Detetermine Number of Sample Sequences}
\usage{
n_sample_sequences(sample.shortreadq)
}
\arguments{
\item{sample.shortreadq}{A ShortRead object containing the sample sequences}
}
\value{
The number of sequences
}
\description{
This function determines the number of sample sequences.
}
