% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{replace_ambiguous_letters}
\alias{replace_ambiguous_letters}
\title{Replace Ambiguous Letters}
\usage{
replace_ambiguous_letters(sample_filetype_FASTQ.character,
  replace_ambiguous_letters.character, sample.shortreadq)
}
\arguments{
\item{sample_filetype_FASTQ.character}{Sample file type (TRUE = fastq, FALSE = fasta)}

\item{replace_ambiguous_letters.character}{Should Ns be replaced? (TRUE = Yes, FALSE = No)}

\item{sample.shortreadq}{A ShortRead object containing the randomly sampled sequences}
}
\value{
A ShortRead object containing the sample sequences with Ns replaced
}
\description{
This function replaces any ambiguous letters (N) in the sample sequence with a random letter (A, T, C, G)
}
