% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced--28 Aug 19.R
\name{sample_file_type}
\alias{sample_file_type}
\title{Determine Sample File Type}
\usage{
sample_file_type(sample.filepath)
}
\arguments{
\item{sample.filepath}{Path to file for sample sequences}
}
\value{
A binary (TRUE is fastq, FALSE is fasta)
}
\description{
This function determines whether sample file type is fastq or fasta.
}
