% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{subsample_sequences}
\alias{subsample_sequences}
\title{Subsample Sequences}
\usage{
subsample_sequences(sample_filtered.shortreadq, n_sample_max.scalar, rand)
}
\arguments{
\item{sample_filtered.shortreadq}{The set of sequencing reads to be subsampled}

\item{n_sample_max.scalar}{The number of reads to be subsampled}

\item{rand}{A seed for the random number generator}
}
\value{
A ShortRead object containing the randomly sampled sequences
}
\description{
This function takes a random subsample of sample sequences.
}
