% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distanced.R
\name{sample_file_type}
\alias{sample_file_type}
\title{Determine Sample File Type}
\usage{
sample_file_type(sample.filepath)
}
\arguments{
\item{sample.filepath}{Path to file for sample sequences}
}
\value{
A logical; TRUE if fastq, FALSE if fasta
}
\description{
This function determines whether sample file type is fastq or fasta.
}
